added support for inline comments
