+++ packages/react/swc/src/utils.rs
@@ -1,0 +1,319 @@
+use regex::Regex;
+use swc_core::{
+    common::{sync::Lazy, Span, DUMMY_SP},
+    ecma::{
+        ast::*,
+        visit::{Visit, VisitWith},
+    },
+};
+
+pub trait FunctionLikeExpr {
+    fn is_regular(&self) -> bool;
+}
+
+impl FunctionLikeExpr for ArrowExpr {
+    fn is_regular(&self) -> bool {
+        !self.is_async && !self.is_generator
+    }
+}
+impl FunctionLikeExpr for FnExpr {
+    fn is_regular(&self) -> bool {
+        self.function.is_regular()
+    }
+}
+impl FunctionLikeExpr for Function {
+    fn is_regular(&self) -> bool {
+        !self.is_async && !self.is_generator
+    }
+}
+impl FunctionLikeExpr for FnDecl {
+    fn is_regular(&self) -> bool {
+        self.function.is_regular()
+    }
+}
+
+fn is_component_name(name: &str) -> bool {
+    static RE: Lazy<Regex> = Lazy::new(|| Regex::new("^[A-Z]").unwrap());
+    RE.is_match(name)
+}
+
+pub trait MaybeComponentName {
+    fn is_component_name(&self) -> bool;
+}
+
+impl MaybeComponentName for Ident {
+    fn is_component_name(&self) -> bool {
+        is_component_name(self.sym.as_str())
+    }
+}
+impl MaybeComponentName for BindingIdent {
+    fn is_component_name(&self) -> bool {
+        self.id.is_component_name()
+    }
+}
+impl MaybeComponentName for Pat {
+    fn is_component_name(&self) -> bool {
+        if let Pat::Ident(id) = self {
+            return id.is_component_name();
+        }
+        false
+    }
+}
+
+pub enum FunctionLike<'a> {
+    Arrow(&'a mut ArrowExpr),
+    Fn(&'a mut FnExpr),
+}
+
+pub fn extract_fn_from_expr<'a>(expr: &'a mut Expr) -> Option<FunctionLike<'a>> {
+    match expr {
+        Expr::Fn(fn_expr) if fn_expr.is_regular() => Some(FunctionLike::Fn(fn_expr)),
+        Expr::Arrow(arrow_expr) if arrow_expr.is_regular() => Some(FunctionLike::Arrow(arrow_expr)),
+        Expr::Call(CallExpr {
+            args,
+            span: _,
+            type_args: _,
+            callee: _,
+        }) => {
+            if let Some(ExprOrSpread {
+                spread: None,
+                expr: first_arg_expr,
+            }) = args.as_mut_slice().take_first_mut()
+            {
+                extract_fn_from_expr(first_arg_expr.unwrap_parens_mut())
+            } else {
+                None
+            }
+        }
+        _ => None,
+    }
+}
+
+pub trait Blockable {
+    fn to_block(&mut self) -> BlockStmt;
+}
+impl Blockable for BlockStmtOrExpr {
+    fn to_block(&mut self) -> BlockStmt {
+        match self {
+            BlockStmtOrExpr::BlockStmt(block) => block.to_owned(),
+            BlockStmtOrExpr::Expr(expr) => BlockStmt {
+                span: DUMMY_SP,
+                stmts: vec![Stmt::Return(ReturnStmt {
+                    span: DUMMY_SP,
+                    arg: Some(expr.clone()),
+                })],
+            },
+        }
+    }
+}
+struct HasJSX {
+    found: bool,
+}
+impl Visit for HasJSX {
+    fn visit_jsx_element(&mut self, _: &JSXElement) {
+        self.found = true;
+    }
+}
+
+struct HasDotValue {
+    found: bool,
+}
+impl Visit for HasDotValue {
+    fn visit_member_expr(&mut self, n: &MemberExpr) {
+        if self.found {
+            return;
+        }
+
+        if match &n.prop {
+            MemberProp::Ident(ident) => ident.sym.as_str() == "value",
+            MemberProp::Computed(ComputedPropName { span: _, expr }) => {
+                if let Expr::Lit(Lit::Str(Str {
+                    span: _,
+                    value,
+                    raw: _,
+                })) = expr.unwrap_parens()
+                {
+                    value.as_str() == "value"
+                } else {
+                    false
+                }
+            }
+            _ => false,
+        } {
+            self.found = true;
+            return;
+        }
+        n.visit_children_with(self);
+    }
+}
+
+fn has_jsx<N>(n: &N) -> bool
+where
+    N: VisitWith<HasJSX>,
+{
+    let mut v = HasJSX { found: false };
+    n.visit_children_with(&mut v);
+    v.found
+}
+fn has_dot_value<N>(n: &N) -> bool
+where
+    N: VisitWith<HasDotValue>,
+{
+    let mut v = HasDotValue { found: false };
+    n.visit_children_with(&mut v);
+    v.found
+}
+
+pub trait Detectable {
+    fn has_jsx(&self) -> bool;
+    fn has_dot_value(&self) -> bool;
+}
+
+impl Detectable for FunctionLike<'_> {
+    fn has_jsx(&self) -> bool {
+        match self {
+            FunctionLike::Arrow(arrow_expr) => has_jsx(*arrow_expr),
+            FunctionLike::Fn(fn_expr) => has_jsx(*fn_expr),
+        }
+    }
+    fn has_dot_value(&self) -> bool {
+        match self {
+            FunctionLike::Arrow(arrow_expr) => has_dot_value(*arrow_expr),
+            FunctionLike::Fn(fn_expr) => has_dot_value(*fn_expr),
+        }
+    }
+}
+impl Detectable for FnDecl {
+    fn has_jsx(&self) -> bool {
+        has_jsx(&self.function)
+    }
+    fn has_dot_value(&self) -> bool {
+        has_dot_value(&self.function)
+    }
+}
+
+pub fn add_import(ident: Ident, source: Str, source_member_ident: Option<Ident>) -> ImportDecl {
+    ImportDecl {
+        span: DUMMY_SP,
+        specifiers: vec![if let Some(source_member_ident) = source_member_ident {
+            ImportSpecifier::Named(ImportNamedSpecifier {
+                span: DUMMY_SP,
+                local: ident,
+                is_type_only: false,
+                imported: Some(ModuleExportName::Ident(source_member_ident)),
+            })
+        } else {
+            ImportSpecifier::Default(ImportDefaultSpecifier {
+                span: DUMMY_SP,
+                local: ident,
+            })
+        }],
+        src: source.into(),
+        type_only: false,
+        with: None,
+    }
+}
+
+pub fn add_require(ident: Ident, source: Str, source_member_ident: Option<Ident>) -> Stmt {
+    Stmt::Decl(Decl::Var(Box::new(VarDecl {
+        span: DUMMY_SP,
+        kind: VarDeclKind::Const,
+        declare: false,
+        decls: vec![VarDeclarator {
+            definite: false,
+            span: DUMMY_SP,
+            name: Pat::Ident(BindingIdent {
+                id: ident,
+                type_ann: None,
+            }),
+            init: {
+                let import_call = Expr::Call(CallExpr {
+                    span: DUMMY_SP,
+                    type_args: None,
+                    callee: Callee::Expr(Box::new(Expr::Ident(Ident {
+                        span: DUMMY_SP,
+                        sym: "require".into(),
+                        optional: false,
+                    }))),
+                    args: vec![ExprOrSpread {
+                        spread: None,
+                        expr: Box::new(Expr::Lit(Lit::Str(source))),
+                    }],
+                });
+
+                if let Some(source_member_ident) = source_member_ident {
+                    Some(Box::new(Expr::Member(MemberExpr {
+                        span: DUMMY_SP,
+                        obj: Box::new(import_call),
+                        prop: MemberProp::Ident(source_member_ident),
+                    })))
+                } else {
+                    Some(Box::new(import_call))
+                }
+            },
+        }],
+    })))
+}
+
+pub trait Spanned {
+    fn get_span(&self) -> &Span;
+}
+
+impl Spanned for Lit {
+    fn get_span(&self) -> &Span {
+        match self {
+            Lit::Bool(lit) => &lit.span,
+            Lit::Num(lit) => &lit.span,
+            Lit::Str(lit) => &lit.span,
+            Lit::BigInt(lit) => &lit.span,
+            Lit::Null(lit) => &lit.span,
+            Lit::Regex(lit) => &lit.span,
+            Lit::JSXText(lit) => &lit.span,
+        }
+    }
+}
+
+impl Spanned for Expr {
+    fn get_span(&self) -> &Span {
+        match self {
+            Expr::Array(lit) => &lit.span,
+            Expr::Arrow(lit) => &lit.span,
+            Expr::Assign(lit) => &lit.span,
+            Expr::Await(lit) => &lit.span,
+            Expr::Bin(lit) => &lit.span,
+            Expr::Call(lit) => &lit.span,
+            Expr::Class(lit) => &lit.class.span,
+            Expr::Cond(lit) => &lit.span,
+            Expr::Fn(lit) => &lit.function.span,
+            Expr::Ident(lit) => &lit.span,
+            Expr::Lit(lit) => lit.get_span(),
+            Expr::Member(lit) => &lit.span,
+            Expr::MetaProp(lit) => &lit.span,
+            Expr::New(lit) => &lit.span,
+            Expr::Object(lit) => &lit.span,
+            Expr::Paren(lit) => &lit.span,
+            Expr::PrivateName(lit) => &lit.span,
+            Expr::Seq(lit) => &lit.span,
+            Expr::TaggedTpl(lit) => &lit.span,
+            Expr::This(lit) => &lit.span,
+            Expr::Tpl(lit) => &lit.span,
+            Expr::Unary(lit) => &lit.span,
+            Expr::Update(lit) => &lit.span,
+            Expr::Yield(lit) => &lit.span,
+            Expr::JSXMember(lit) => &lit.prop.span,
+            Expr::JSXNamespacedName(lit) => &lit.ns.span,
+            Expr::JSXEmpty(lit) => &lit.span,
+            Expr::JSXElement(lit) => &lit.span,
+            Expr::JSXFragment(lit) => &lit.span,
+            Expr::TsTypeAssertion(lit) => &lit.span,
+            Expr::TsConstAssertion(lit) => &lit.span,
+            Expr::TsNonNull(lit) => &lit.span,
+            Expr::TsAs(lit) => &lit.span,
+            Expr::TsInstantiation(lit) => &lit.span,
+            Expr::TsSatisfies(lit) => &lit.span,
+            Expr::Invalid(it) => &it.span,
+            Expr::SuperProp(it) => &it.span,
+            Expr::OptChain(it) => &it.span,
+        }
+    }
+}
